Implemented support for Slash commands

+++ src/main/kotlin/com/pb/scheduling/data/ScheduleRequest.kt
@@ -4,6 +4,6 @@ import dev.kord.core.entity.User
 
 data class ScheduleRequest(
     val user: User,
-    val delayInMinutes: Int,
+    val delayInMinutes: Long,
     val action: suspend () -> Unit,
 )
